Move non-api tests into app package
